% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcd.R
\docType{data}
\name{abcd_specs_examples}
\alias{abcd_specs_examples}
\alias{abcd_specs_complete}
\alias{abcd_specs_without_conditions}
\alias{abcd_specs_single_po_without_conditions}
\alias{abcd_specs_dutch_xtc}
\alias{abcd_specification_example_xtc}
\alias{abcd_specification_empty}
\title{Simple example datasets for ABCDs}
\format{For \code{abcd_specs_complete}, a data frame with 7 variables and 7 rows;
for \code{abcd_specs_without_conditions}, a data frame with 6 variables and 7 rows;
for \code{abcd_specs_single_po_without_conditions}, a data frame with 5 variables and 4 rows;
for \code{abcd_specification_example_xtc} and \code{abcd_specs_dutch_xtc},
a data frame with 7 variables and 5 rows' and
for \code{abcd_specification_empty}, a data frame with 7 variables and 1 row.}
\usage{
data(abcd_specs_complete)

data(abcd_specs_without_conditions)

data(abcd_specs_single_po_without_conditions)

data(abcd_specification_example_xtc)

data(abcd_specs_dutch_xtc)

data(abcd_specification_empty)
}
\description{
This are three (nested) datasets illustrating the logic model of change for
a simple condom use intervention in a way that can be visualised using
the \link{abcd} function. The full dataset is \code{abcd_specs_full}, a subset
that does not explicitly include the conditions for effectiveness
(instead showing letters that can then be explained in, for example,
the manuscript text) is called \code{abcd_specs_without_conditions}, and
a version that only contains the information about one sub-behavior
(performance objective) is available as
\code{abcd_specs_single_po_without_conditions}. The variables in the full
dataset are:
}
\details{
\itemize{
\item \verb{Behavior Change Principles}: The behavior change principles (BCPs), also known as methods for behavior change or 'behavior change techniques' (BCTs), that describe the psychological principles that are assumed to realise the change in the (sub-)determinants.
\item \verb{Conditions for effectiveness\\\\n(e.g. parameters for use)}: The conditions for effectiveness that describe the constraints and considerations taken into account in the translation of the BCPs to practical applications for the relevant target population, context, culture, etc.
\item \code{Applications}: The applications of these BCPs. Where the BCPs describe theoretical principles, the applications are more or less tangible intervention elements.
\item \verb{Sub-determinants\\\\n(e.g. beliefs; can be formulated as Change Objectives)}: The specific aspects of teh target population's psychology that are targeted by the BCPs (e.g. beliefs, or in Intervention Mapping vocabulary, Change Objectives).
\item \code{Determinants}: The determinants, psychological constructs, that the targeted sub-determinants are a part of, and that together predict the Performance Objectives (sub-behaviors).
\item \verb{Performance Objectives}: Explicitly defined sub-behaviors at a level of specificity that distinguishes them from other sub-behaviors, and that together form the target behavior.
\item \verb{Target Behavior}: The ultimate target behavior, usually defined at a relatively general level.
}

In addition to these three datasets, a Dutch example specification
is included named \code{abcd_specs_dutch_xtc}, and the same in English as \code{abcd_specification_example_xtc}.

Finally, \code{abcd_specification_empty} is an empty 'template' ABCD matrix.
}
\keyword{data}
